Add Cisco IOS XE driver
On Cisco IOS XE the configration is immediately "commited" when a CLI
command is accepted in a configuration session. This means that simply
exiting an active configuration session is not sufficient to abort
configuration because of an error. We use the "config archive" feature
to instruct the device to keep track of CLI commands entered.

The driver creates an archive checkpoint before each configuration
change and maintains the history in a stack. This is then used for both
automatic (error recovery) and user-triggered rollback operations.

The config archive must be enabled on the device prior to starting the
configuration. As it stands today, this is the responsibility of the
user. The configuration is:

  archive
   log config
    logging enable
  path bootflash:archive